requirements.txt
requests
beautifulsoup4
python-dotenv
fastapi
streamlit
google-generativeai
pathlib
uvicorn